# storage/swagger/http.json
{"openapi":"3.0.0","paths":{"\/login":{"post":{"tags":["注册登录"],"summary":"小程序登录","operationId":"af704524b829a3c6cea821006f70e2b0","requestBody":{"content":{"application\/json":{"schema":{"properties":{"code":{"description":"微信授权码","type":"string","rules":"string"}},"type":"object"}}},"_content":{"properties":{"code":{"description":"微信授权码","type":"string","rules":"string"}},"type":"object"}},"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/LoginSchema"}}}}}}},"\/task":{"get":{"tags":["任务管理"],"summary":"任务列表","operationId":"c144ae4848ead5b67dd693df33b11214","parameters":[{"parameter":"offset","in":"query","description":"偏移量","rules":"integer"},{"parameter":"limit","in":"query","description":"单页条数","rules":"integer"}],"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/TaskListSchema"}}}}}}},"\/task\/{id:\\d+}":{"get":{"tags":["任务管理"],"summary":"任务详情","operationId":"2fb5cb8f1debe76f2c49d7e1016fe11e","parameters":[{"name":"id","in":"path","description":"任务ID"}],"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/TaskSchema"}}}}}},"post":{"tags":["任务管理"],"summary":"更新任务","operationId":"e84696e4af1e5267402d74c2f9d12740","parameters":[{"name":"id","in":"path","description":"任务ID 0新增 >0修改"}],"requestBody":{"content":{"application\/json":{"schema":{"properties":{"name":{"description":"任务名","type":"string","rules":"required|string"},"summary":{"description":"任务描述","type":"string","rules":"string"}},"type":"object"}}},"_content":{"properties":{"name":{"description":"任务名","type":"string","rules":"required|string"},"summary":{"description":"任务描述","type":"string","rules":"string"}},"type":"object"}},"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/SavedSchema"}}}}}}},"\/task\/{id:\\d+}\/delete":{"post":{"tags":["任务管理"],"summary":"删除任务","operationId":"72b522f862adf82aebb142195fcd2a3c","parameters":[{"name":"id","in":"path","description":"任务ID"}],"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/SavedSchema"}}}}}}},"\/task\/{id:\\d+}\/top":{"post":{"tags":["任务管理"],"summary":"置顶任务","operationId":"5b42342a1d6f40b8b70c76e1cf8bab3c","parameters":[{"name":"id","in":"path","description":"任务ID"}],"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/SavedSchema"}}}}}}},"\/task-item\/{id:\\d+}":{"get":{"tags":["任务记录管理"],"summary":"任务记录详情","operationId":"8954e158993a75d757a409d332ba492a","parameters":[{"name":"id","in":"path","description":"任务ID"}],"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/TaskItemSchema"}}}}}},"post":{"tags":["任务记录管理"],"summary":"更新记录","operationId":"544e499f90af4c2924c0eb0fe5876494","parameters":[{"name":"id","in":"path","description":"任务记录ID"}],"requestBody":{"content":{"application\/json":{"schema":{"properties":{"task_id":{"description":"任务ID","type":"integer","rules":"required|integer"},"value":{"description":"数值","type":"string","rules":"required|numeric"},"comment":{"description":"备注","type":"string","rules":"string"}},"type":"object"}}},"_content":{"properties":{"task_id":{"description":"任务ID","type":"integer","rules":"required|integer"},"value":{"description":"数值","type":"string","rules":"required|numeric"},"comment":{"description":"备注","type":"string","rules":"string"}},"type":"object"}},"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/SavedSchema"}}}}}}},"\/task-item":{"get":{"tags":["任务记录管理"],"summary":"任务记录列表","operationId":"180956d74ed73e78c5fb2b01d14fc926","parameters":[{"parameter":"task_id","in":"query","description":"任务ID","rules":"required|integer"},{"parameter":"offset","in":"query","description":"偏移量","rules":"integer"},{"parameter":"limit","in":"query","description":"单页条数","rules":"integer"}],"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/TaskItemListSchema"}}}}}}},"\/task-item\/chart":{"get":{"tags":["任务记录管理"],"summary":"任务记录图表","operationId":"f00c206d4a2c0167b5b84bb7df9f3624","parameters":[{"parameter":"task_id","in":"query","description":"任务ID","rules":"required|integer"}],"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/ChartSchema"}}}}}}},"\/task-item\/{id:\\d+}\/delete":{"post":{"tags":["任务记录管理"],"summary":"删除任务记录","operationId":"1236b57e58c06827da65bfde5a8c5cc4","parameters":[{"name":"id","in":"path","description":"任务ID"}],"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/SavedSchema"}}}}}}},"\/ys-player\/create":{"post":{"tags":["原神练度管理"],"summary":"添加原神账号","operationId":"8f99a8d0c8ecc26497b7e4f1f6ea4851","requestBody":{"content":{"application\/json":{"schema":{"properties":{"uid":{"description":"原神 UID","type":"integer","rules":"required|integer"},"comment":{"description":"备注","type":"string","rules":"string"}},"type":"object"}}},"_content":{"properties":{"uid":{"description":"原神 UID","type":"integer","rules":"required|integer"},"comment":{"description":"备注","type":"string","rules":"string"}},"type":"object"}},"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/SavedSchema"}}}}}}},"\/ys-player":{"get":{"tags":["原神练度管理"],"summary":"原神账号列表","operationId":"6f5f8b797ccede7ccb627e17cc613fda","parameters":[{"parameter":"offset","in":"query","description":"偏移量","rules":"integer"},{"parameter":"limit","in":"query","description":"单页条数","rules":"integer"}],"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/YsPlayerListSchema"}}}}}}},"\/ys-player\/{id:\\d+}":{"get":{"tags":["原神练度管理"],"summary":"原神账号详情","operationId":"e0bb38c0b0e2e52f138d6bd42306abf8","parameters":[{"name":"id","in":"path","description":"原神账号 ID"}],"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/YsPlayerListSchema"}}}}}},"delete":{"tags":["原神练度管理"],"summary":"原神账号详情","operationId":"e882a53d2e2a52c1b15bbe0de3aff224","parameters":[{"name":"id","in":"path","description":"原神账号 ID"}],"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/SavedSchema"}}}}}}},"\/ys-roler\/{id:\\d+}":{"get":{"tags":["原神练度管理"],"summary":"查看角色详情","operationId":"413478e8020e6646b641e77fe532fe96","parameters":[{"name":"id","in":"path","description":"原神账号 ID"}],"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/YsRolerSchema"}}}}}}},"\/ys-roler\/target":{"post":{"tags":["原神练度管理"],"summary":"设置角色目标","operationId":"1e1b8034681b006ad86d1da7eb2e5da8","requestBody":{"content":{"application\/json":{"schema":{"properties":{"roler_id":{"description":"原神角色 ID","type":"integer","rules":"required|integer"},"level":{"description":"等级","type":"integer","rules":"required|integer"},"hp":{"description":"生命值","type":"integer","rules":"required|integer"},"attack":{"description":"攻击力","type":"integer","rules":"required|integer"},"defend":{"description":"防御力","type":"integer","rules":"required|integer"},"element":{"description":"元素精通","type":"integer","rules":"required|integer"},"crit":{"description":"暴击率","type":"number","format":"float","rules":"required|float"},"crit_dmg":{"description":"暴击伤害","type":"number","format":"float","rules":"required|float"},"recharge":{"description":"充能效率","type":"number","format":"float","rules":"required|float"},"heal":{"description":"属性伤害加成","type":"number","format":"float","rules":"required|float"}},"type":"object"}}},"_content":{"properties":{"roler_id":{"description":"原神角色 ID","type":"integer","rules":"required|integer"},"level":{"description":"等级","type":"integer","rules":"required|integer"},"hp":{"description":"生命值","type":"integer","rules":"required|integer"},"attack":{"description":"攻击力","type":"integer","rules":"required|integer"},"defend":{"description":"防御力","type":"integer","rules":"required|integer"},"element":{"description":"元素精通","type":"integer","rules":"required|integer"},"crit":{"description":"暴击率","type":"number","format":"float","rules":"required|float"},"crit_dmg":{"description":"暴击伤害","type":"number","format":"float","rules":"required|float"},"recharge":{"description":"充能效率","type":"number","format":"float","rules":"required|float"},"heal":{"description":"属性伤害加成","type":"number","format":"float","rules":"required|float"}},"type":"object"}},"responses":{"200":{"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/SavedSchema"}}}}}}}},"components":{"schemas":{"ChartMetaSchema":{"title":"图表元数据","properties":{"key":{"title":"x坐标","type":"string","rules":null},"value":{"title":"y坐标","type":"string","rules":null},"value_str":{"title":"y坐标","type":"string","nullable":true,"rules":null}},"type":"object"},"ChartSchema":{"title":"图表数据","properties":{"name":{"title":"图表名","type":"string","rules":null},"values":{"title":"图表数据","type":"array","items":{"$ref":"#\/components\/schemas\/ChartMetaSchema"},"rules":null}},"type":"object"},"LoginSchema":{"title":"登录返回信息","properties":{"token":{"title":"token","type":"string","rules":null}},"type":"object"},"SavedSchema":{"title":"修改结果","properties":{"saved":{"title":"是否保存成功","type":"boolean","rules":null}},"type":"object"},"TaskItemListSchema":{"title":"任务记录列表","properties":{"count":{"title":"总数","type":"integer","rules":null},"list":{"title":"列表","type":"integer","rules":null}},"type":"object"},"TaskItemSchema":{"title":"任务记录详情","properties":{"id":{"title":"","type":"integer","rules":null},"user_id":{"title":"用户ID","type":"integer","rules":null},"task_id":{"title":"任务ID","type":"integer","rules":null},"date":{"title":"日期","type":"mixed","rules":null},"value":{"title":"数值","type":"mixed","rules":null},"comment":{"title":"备注","type":"string","rules":null},"created_at":{"title":"创建时间","type":"string","rules":null},"updated_at":{"title":"更新时间","type":"string","rules":null}},"type":"object"},"TaskListSchema":{"title":"任务列表","properties":{"count":{"title":"总数","type":"integer","rules":null},"list":{"title":"列表","type":"integer","rules":null}},"type":"object"},"TaskSchema":{"title":"任务详情","properties":{"id":{"title":"","type":"integer","rules":null},"user_id":{"title":"用户ID","type":"integer","rules":null},"name":{"title":"任务标题","type":"string","rules":null},"summary":{"title":"备注","type":"string","rules":null},"is_deleted":{"title":"0未删除 1已删除","type":"integer","rules":null},"created_at":{"title":"创建时间","type":"string","rules":null},"updated_at":{"title":"更新时间","type":"string","rules":null}},"type":"object"},"YsPlayerListSchema":{"title":"原神账号列表","properties":{"count":{"title":"总数","type":"integer","rules":null},"list":{"$ref":"#\/components\/schemas\/YsPlayerSchema"}},"type":"object"},"YsPlayerSchema":{"title":"原神账号详情","properties":{"id":{"title":"","type":"integer","rules":null},"user_id":{"title":"用户 ID","type":"integer","rules":null},"uid":{"title":"原神 UID","type":"integer","rules":null},"comment":{"title":"备注","type":"string","rules":null},"rolers":{"title":"角色列表","type":"array","items":{"$ref":"#\/components\/schemas\/YsRolerSchema"},"rules":null}},"type":"object"},"YsRolerSchema":{"title":"原神角色详情","properties":{"id":{"title":"","type":"integer","rules":null},"user_id":{"title":"用户 ID","type":"integer","rules":null},"uid":{"title":"原神 UID","type":"integer","rules":null},"role":{"title":"角色名","type":"string","rules":null},"role_img":{"title":"角色头像","type":"string","rules":null},"level":{"title":"角色等级","type":"integer","rules":null},"artifacts_sum_point":{"title":"圣遗物分值","type":"integer","rules":null},"hp":{"title":"生命值","type":"integer","rules":null},"attack":{"title":"攻击力","type":"integer","rules":null},"defend":{"title":"防御力","type":"integer","rules":null},"element":{"title":"元素精通","type":"integer","rules":null},"crit":{"title":"暴击率","type":"string","rules":null},"crit_dmg":{"title":"暴击伤害","type":"string","rules":null},"recharge":{"title":"充能效率","type":"string","rules":null},"heal":{"title":"属性伤害加成","type":"string","rules":null},"target":{"$ref":"#\/components\/schemas\/YsRolerTargetSchema"}},"type":"object"},"YsRolerTargetSchema":{"title":"原神角色目标详情","properties":{"id":{"title":"原神角色ID","type":"integer","rules":null},"level":{"title":"角色等级","type":"integer","rules":null},"hp":{"title":"生命值","type":"integer","rules":null},"attack":{"title":"攻击力","type":"integer","rules":null},"defend":{"title":"防御力","type":"integer","rules":null},"element":{"title":"元素精通","type":"integer","rules":null},"crit":{"title":"暴击率","type":"string","rules":null},"crit_dmg":{"title":"暴击伤害","type":"string","rules":null},"recharge":{"title":"充能效率","type":"string","rules":null},"heal":{"title":"属性伤害加成","type":"string","rules":null}},"type":"object"}}}}